feat: show filters only on document view +refactor DocumentLibrary to only switch between 2 views
