+++ frontend/.vscode/settings.json
@@ -4,7 +4,5 @@
   "editor.codeActionsOnSave": {
     "source.organizeImports": "always"
   },
-  "i18n-ally.localesPaths": [
-    "src/locales",
-  ],
+  "i18n-ally.localesPaths": ["src/locales"]
 }
